Use nodeunit test framework for unit tests
After reviewing several test frameworks available for node, I came to
conclusion that [nodeunit](https://github.com/caolan/nodeunit) seems to
be the most sane one. Convert existing unit test into nodeunit
compatible and add nodeunit into dev dependencies.
